Moved some of the newly integrated 3rd party code around
