+++ src/Hate/Graphics/Backend/Compat/Types.hs
@@ -1,7 +1,7 @@
 module Hate.Graphics.Backend.Compat.Types where
 
 import Hate.Graphics.Types
-import Hate.Graphics.Pipeline
+import Hate.Graphics.Backend.Common.Pipeline
 
 -- supposedly needs more vertex streams and pipelines in the future
 data BackendCompat = BackendCompat {
