+++ src/Hate/Graphics/Backend/Modern/Types.hs
@@ -1,7 +1,7 @@
 module Hate.Graphics.Backend.Modern.Types where
 
 import Hate.Graphics.Types
-import Hate.Graphics.Pipeline
+import Hate.Graphics.Backend.Common.Pipeline
 
 -- supposedly needs more vertex streams and pipelines in the future
 data BackendModern = BackendModern {
